VD.ah2: check WinExist before WinGetMinMax
On Windows 11, I'm getting some errors on WinGetMinMax
that window doesn't exist, I think because we still get
the hwnd from the old desktop which doesn't exist on the
new desktop, maybe some form of race condition.
Adding the check for WinExist avoids the error, and it
works as expected with the window under activated.